Add ui package for reusable components (#1916)
* Add ui package for reusable components

* Add fallback

* Type fixes

* Type fixes
