+++ apps/web/next.config.js
@@ -1,4 +1,8 @@
-const withTM = require("@vercel/edge-functions-ui/transpile")(["@calcom/lib", "@calcom/prisma"]);
+const withTM = require("@vercel/edge-functions-ui/transpile")([
+  "@calcom/lib",
+  "@calcom/prisma",
+  "@calcom/ui",
+]);
 const { i18n } = require("./next-i18next.config");
 
 // So we can test deploy previews preview
