+++ apps/web/package.json
@@ -29,6 +29,7 @@
     "@calcom/lib": "*",
     "@calcom/prisma": "*",
     "@calcom/tsconfig": "*",
+    "@calcom/ui": "*",
     "@daily-co/daily-js": "^0.21.0",
     "@headlessui/react": "^1.4.2",
     "@heroicons/react": "^1.0.5",
