+++ packages/lib/classNames.ts
@@ -1,0 +1,3 @@
+export default function classNames(...classes: unknown[]) {
+  return classes.filter(Boolean).join(" ");
+}
