+++ packages/ui/index.tsx
@@ -1,0 +1,3 @@
+import * as React from "react";
+
+export * from "./Button";
