+++ packages/ui/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "@calcom/ui",
+  "version": "0.0.0",
+  "main": "./index.tsx",
+  "types": "./index.tsx",
+  "license": "MIT",
+  "devDependencies": {
+    "@calcom/config": "*",
+    "@calcom/lib": "*",
+    "@calcom/tsconfig": "*",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "typescript": "^4.5.3"
+  }
+}
