+++ packages/ui/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@calcom/tsconfig/react-library.json",
+  "include": ["."],
+  "exclude": ["dist", "build", "node_modules"]
+}
